Support VECTOR for MariaDB 11.7.

+++ tests/IntegrationTests/DataTypesFixture.cs
@@ -245,6 +245,10 @@ values
 
 		if (AppConfig.SupportedFeatures.HasFlag(ServerFeatures.Vector))
 		{
+			// create a helper function for MariaDB 11.7+
+			if (Connection.ServerVersion.StartsWith("11.7.", StringComparison.Ordinal))
+				Connection.Execute("create function if not exists STRING_TO_VECTOR(s text) returns vector(3) deterministic return Vec_FromText(s);");
+
 			Connection.Execute("""
 				drop table if exists datatypes_vector;
 				create table datatypes_vector (
@@ -257,7 +261,7 @@ values
 					(STRING_TO_VECTOR('[0, 0, 0]')),
 					(STRING_TO_VECTOR('[1, 1, 1]')),
 					(STRING_TO_VECTOR('[1, 2, 3]')),
-					(STRING_TO_VECTOR('[3.40282347E+38, 3.40282347E+38, 3.40282347E+38]'));
+					(STRING_TO_VECTOR('[-1, -1, -1]'));
 				""");
 		}
 		Connection.Close();
